# [wasm] Optimize output from threaded test runner