--- a/src/Microsoft.DotNet.XHarness.TestRunners.Common/TestRunner.cs
+++ b/src/Microsoft.DotNet.XHarness.TestRunners.Common/TestRunner.cs
@@ -144,8 +144,6 @@ public abstract class TestRunner
             return;
         }
 
-        Console.WriteLine($"MF LogFailureSummary {FailureInfos.Count}");
-
         OnInfo("Failed tests:");
         for (int i = 1; i <= FailureInfos.Count; i++)
         {