--- a/src/Microsoft.DotNet.XHarness.TestRunners.Xunit/WasmXmlResultWriter.cs
+++ b/src/Microsoft.DotNet.XHarness.TestRunners.Xunit/WasmXmlResultWriter.cs
@@ -1,25 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Xml.Linq;
 
-namespace Microsoft.DotNet.XHarness.TestRunners.Xunit
-{
-    public class WasmXmlResultWriter
-    {
-        public static void WriteOnSingleLine(XElement assembliesElement)
-        {
-            using var ms = new MemoryStream();
-            assembliesElement.Save(ms);
-            ms.TryGetBuffer(out var bytes);
-            var base64 = Convert.ToBase64String(bytes, Base64FormattingOptions.None);
-            Console.WriteLine($"STARTRESULTXML {bytes.Count} {base64} ENDRESULTXML");
-            Console.WriteLine($"Finished writing {bytes.Count} bytes of RESULTXML");
-        }
+#nullable enable
 
-        public static string EscapeNewLines(string message) => message.Replace("\r", "\\r").Replace("\n", "\\n");
+namespace Microsoft.DotNet.XHarness.TestRunners.Xunit;
+
+public class WasmXmlResultWriter
+{
+    public static void WriteOnSingleLine(XElement assembliesElement)
+    {
+        using var ms = new MemoryStream();
+        assembliesElement.Save(ms);
+        ms.TryGetBuffer(out var bytes);
+        var base64 = Convert.ToBase64String(bytes, Base64FormattingOptions.None);
+        Console.WriteLine($"STARTRESULTXML {bytes.Count} {base64} ENDRESULTXML");
+        Console.WriteLine($"Finished writing {bytes.Count} bytes of RESULTXML");
     }
 }